Adding test to Prime number function
